Remove duplicate createItem function
Always import createItem function from dashboards comands module.
