+++ gsa/src/gmp/commands/__tests__/dashboards.js
@@ -22,6 +22,7 @@
  */
 import {
   DEFAULT_ROW_HEIGHT,
+  createItem,
   createRow,
 } from '../dashboards';
 
@@ -48,3 +49,26 @@ describe('createRow tests', () => {
   });
 
 });
+
+describe('createItem tests', () => {
+
+  test('should create a new item with empty props', () => {
+    const uuid = jest.fn().mockReturnValue(1);
+    expect(createItem(undefined, uuid)).toEqual({
+      id: 1,
+    });
+    expect(uuid).toHaveBeenCalled();
+  });
+
+  test('should create a new item with props', () => {
+    const uuid = jest.fn().mockReturnValue(1);
+    expect(createItem({foo: 'bar'}, uuid)).toEqual({
+      id: 1,
+      foo: 'bar',
+    });
+    expect(uuid).toHaveBeenCalled();
+  });
+
+});
+
+// vim: set ts=2 sw=2 tw=80:
